@@ -1,51 +1,34 @@
-<template>
-  <Nav />
-  <Main />
-  <NotificationList />
-  <Loader v-if="!ready" />
-</template>
-<script>
+<script setup>
+import { useIndexStore } from "./stores/index.js";
 import Nav from "./components/Nav.vue";
 import Main from "./components/Main.vue";
-import NotificationList from "./components/NotificationList.vue";
-import Loader from "./components/Loader.vue";
-export default {
-  name: "App",
-  components: {
-    Nav,
-    Main,
-    NotificationList,
-    Loader
-  },
-  computed: {
-    ready() {
-      return (
-        this.$store.state.process.remote === "Done." &&
-        this.$store.state.process.local === "Done."
-      );
-    }
-  },
-  mounted() {
-    this.$store.dispatch("init");
-  }
-};
+import Activities from "./components/Activities.vue";
+import { computed } from "vue";
+
+const store = useIndexStore();
+
+const activities = computed(() => {
+  return store.activities.length > 0;
+});
+store.addActivity("Initializing IndexedDB.");
+store
+  .init()
+  .then(() => {
+    store.removeActivity("Initializing IndexedDB.");
+    store.addActivity("Loading local data.");
+    Promise.allSettled([
+      store.loadLocalDirectories(),
+      store.loadLocalFiles()
+    ]).then((results) => {
+      store.removeActivity("Loading local data.");
+      return results;
+    });
+  })
+  .catch((error) => console.error(error));
 </script>
 
-<style lang="sass">
-*
-  box-sizing: border-box
-  padding: 0
-  margin: 0
-  font-family: Segoe UI, sans serif
-body
-  min-height: 100vh
-  background: rgba(65,184,131,1)
-  @media (prefers-color-scheme: dark)
-    background: rgba(53,73,94,1)
-  *
-    color: rgba(53,73,94,1)
-    color: rgba(255,255,255,1)
-    @media (prefers-color-scheme: dark)
-      color: rgba(65,184,131,1)
-      color: rgba(255,255,255,1)
-</style>
+<template>
+  <Nav></Nav>
+  <Main></Main>
+  <Activities v-if="activities"></Activities>
+</template>
