@@ -1,11 +1,7 @@
+<script setup></script>
+
 <template>
   <main>
     <router-view></router-view>
   </main>
 </template>
-
-<script>
-export default {
-  name: "Main"
-};
-</script>
