@@ -1,8 +1,8 @@
 import { createApp } from "vue";
+import { createPinia } from "pinia";
+
 import App from "./App.vue";
-import "./registerServiceWorker";
 import router from "./router";
-import store from "./store";
-// import "@fortawesome/fontawesome-free";
-require("@/assets/styles.sass");
-createApp(App).use(store).use(router).mount("body");
+import "virtual:windi.css";
+
+createApp(App).use(createPinia()).use(router).mount("#app");
